+++ Server.Core/Entities/Ingredient.cs
@@ -7,8 +7,8 @@ namespace server.Models
     public class Ingredient : BaseEntity
     {
         public string Name { get; set; }   
-        public int PurchaseQuantity { get; set; }
-        public float PurchasePrice { get; set; }
+        public decimal PurchaseQuantity { get; set; }
+        public decimal PurchasePrice { get; set; }
         public UnitEnum PurchaseUnit { get; set; }   
         public List<RecipeIngredients> RecipeIngredients { get; set; }
